moved ExtendedProfile to mocks for later usage
