+++ packages/io-auth-n-identity-domain/src/domain/__mocks__/profile.mock.ts
@@ -4,14 +4,58 @@ import {
   NonEmptyString,
 } from "@pagopa/hexagonal-core";
 
-import { ExtendedProfileSchema, ServicePreferencesSettingsSchema } from "../entities/profile.entity.js";
 import {
   IsInboxEnabledSchema,
   IsWebhookEnabledSchema,
   PreferredLanguageSchema,
   PreferredLanguagesEnum,
   ServicesPreferencesModeEnum,
+  ServicesPreferencesModeSchema,
+  IsEmailEnabledSchema,
+  IsEmailValidatedSchema,
+  IsTestProfileSchema,
+  BlockedInboxOrChannelsSchema,
+  PreferredLanguagesSchema,
+  PushNotificationsContentTypeSchema,
+  AcceptedTosVersionSchema,
+  ReminderStatusSchema,
+  AppVersionSchema,
 } from "../value-objects/profile/profile.value-object.js";
+import { EmailAddressSchema } from "@pagopa/hexagonal-core";
+
+import { z } from "zod";
+
+// NOTE: the following types are untested.
+// Please verify and move those inside a dedicated profile entity when needed
+const ServicePreferencesSettingsSchema = z.object({
+  mode: ServicesPreferencesModeSchema,
+});
+type ServicePreferencesSettingsSchema = z.infer<
+  typeof ServicePreferencesSettingsSchema
+>;
+
+const ExtendedProfileBase = z.object({
+  is_email_enabled: IsEmailEnabledSchema,
+  is_email_validated: IsEmailValidatedSchema,
+  is_email_already_taken: z.boolean().default(false),
+  is_inbox_enabled: IsInboxEnabledSchema,
+  is_webhook_enabled: IsWebhookEnabledSchema,
+  service_preferences_settings: ServicePreferencesSettingsSchema,
+  version: z.number().int(),
+});
+
+const ExtendedProfileSchema = ExtendedProfileBase.extend({
+  email: EmailAddressSchema.optional(),
+  blocked_inbox_or_channels: BlockedInboxOrChannelsSchema.optional(),
+  preferred_languages: PreferredLanguagesSchema.optional(),
+  accepted_tos_version: AcceptedTosVersionSchema.optional(),
+  reminder_status: ReminderStatusSchema.optional(),
+  is_test_profile: IsTestProfileSchema.optional(),
+  last_app_version: AppVersionSchema.optional(),
+  push_notifications_content_type:
+    PushNotificationsContentTypeSchema.optional(),
+});
+type ExtendedProfileSchema = z.infer<typeof ExtendedProfileSchema>;
 
 export const aCustomEmailAddress = "custom-email@example.com" as EmailAddress;
